+++ db/migrate/20161125141914_create_join_table.rb
@@ -1,0 +1,8 @@
+class CreateJoinTable < ActiveRecord::Migration
+  def change
+    create_join_table :tickets, :users do |t|
+      # t.index [:ticket_id, :user_id]
+      # t.index [:user_id, :ticket_id]
+    end
+  end
+end